soapy: switch to native CF32 and use smaller input buffer
This lets Soapy handle conversion and scaling and simplifies the sample
processing routine. We also don't need a huge buffer since we are doing
incremental processing and Soapy will typically limit the actual used
size when used in network mode.

As an added bonus it also fixes the reported message level which is now
in line with native rtlsdr.