msk: use ternary op to improve compiler optimization
for the if (ch->MskS & 1) block, using a ternary op enables the compiler
to efficiently optimize the sign adjustment and real/imag part assignment
using (on x86) a single branch instead of 5.
for the if (ch->MskS & 2) block, using a ternary inside the function
call reduces code size by having only one inlined copy and allows further
integration with the results of the previous if block.

this also makes the code slightly more compact/readable