msk: do not divide power by 4
As dm_buffer from SDR input only contains positive values (cabsf()),
i.e. half of the power scale (just like the matched filter), this
normalization skews the actual dB scale readout and prevents reporting
clipping (i.e. overloads).

This change has been matched with a readout of an audio dump of a live
SDR capture normalized in an audio processing software: the reported
value of the normalized file is '-0.0' as expected, vs '-6.0' with the
previous code. The lowest reported values for SDR input are now also
more in line with the expected SNR.

In order to keep soundfile input values consistent with this change, the
audio samples are "normalized" to half scale. That means that audio
files generated through DEBUG will appear 6dB quieter than the original
signal (i.e. the readback will report the same value as was reported for
SDR input *before* this patch). The rationale is to show "true" values
for "standard" audio input.