msk/acars: add support for noise floor estimator
This introduces a basic, low cpu-cost NF estimator which works by using
an exponential moving average over the magnitude of the received signal
updated only outside of ACARS message processing.

The computational overhead of this extra estimator is virtually nil, as
the signal power estimator has been reworked to use the same input as
the noise floor estimator.

The known limitation is that this estimate will be skewed in case of a
large burst of ACARS messages as the average will no longer converge
towards the noisefloor. To be improved on later.

Fixes: https://github.com/TLeconte/acarsdec/issues/84